+++ app/src/main/java/com/bongtu/baekseo/presentation/setting/SettingScreen.kt
@@ -39,6 +39,7 @@ import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.flowWithLifecycle
@@ -74,18 +75,16 @@ import com.bongtu.baekseo.core.util.UrlConstant
 import com.bongtu.baekseo.core.util.excludeTop
 import com.bongtu.baekseo.core.util.noRippleClickable
 import com.bongtu.baekseo.core.util.openUrl
-import com.bongtu.baekseo.presentation.setting.SettingContract.SettingSideEffect
-import com.bongtu.baekseo.presentation.setting.SettingContract.SettingSideEffect.MainSideEffect.RestartApp
+import com.bongtu.baekseo.presentation.setting.SettingContract.SettingSideEffect.RestartApp
 import com.bongtu.baekseo.presentation.setting.SettingContract.SettingUiState
-import kotlinx.coroutines.flow.filterIsInstance
 
 @Composable
 fun SettingRoute(
-    navigateToEditProfile: () -> Unit,
+    navigateToEditProfile: (String, String, IncomeType) -> Unit,
     navigateToWithdraw: () -> Unit,
     onRestartApp: (Boolean) -> Unit,
-    viewModel: SettingViewModel,
     modifier: Modifier = Modifier,
+    viewModel: SettingViewModel = hiltViewModel(),
 ) {
     val uiState by viewModel.uiState.collectAsStateWithLifecycle()
     val lifecycleOwner = LocalLifecycleOwner.current
@@ -97,7 +96,6 @@ fun SettingRoute(
 
     LaunchedEffect(viewModel.sideEffect, lifecycleOwner) {
         viewModel.sideEffect.flowWithLifecycle(lifecycle = lifecycleOwner.lifecycle)
-            .filterIsInstance<SettingSideEffect.MainSideEffect>()
             .collect { sideEffect ->
                 when (sideEffect) {
                     RestartApp -> onRestartApp(false)
@@ -107,7 +105,13 @@ fun SettingRoute(
 
     SettingScreen(
         uiState = uiState,
-        navigateToEditProfile = navigateToEditProfile,
+        navigateToEditProfile = {
+            navigateToEditProfile(
+                uiState.userName,
+                uiState.userBirth,
+                uiState.userIncome,
+            )
+        },
         navigateToWithdraw = navigateToWithdraw,
         onLogoutClick = viewModel::logout,
         onInquiryClick = { context.openUrl(UrlConstant.INQUIRY_URL) },
